more view updates and first player notification - added notification of the first player - tried to improve synchronization - added controls server side - adding more updates to the client - now the order of the players in multiplayer are chosen randomly

+++ src/main/java/it/polimi/ingsw/common/Events/DeleteProductionEvent.java
@@ -17,17 +17,30 @@ public class DeleteProductionEvent extends Event {
     public void handle(Object playerObj) {
         HumanPlayer player = ((HumanPlayer) playerObj);
 
+        // returning a fail event if it's not the turn of the player
+        if (!player.isPlaying()) {
+            player.getGameClientHandler().sendEvent(new FailEvent("Can't do this action, it's not your turn!"));
+            return;
+        }
+
+        // returning a fail event if the player already did a main action
         if (player.isActionDone()) {
             player.getGameClientHandler().sendEvent(new FailEvent("You already did a main action in this round!"));
             return;
         }
 
-        player.deleteProduction(
-                new MakePlayerChoose<>(
-                        "Choose the production you want to delete",
-                        player.getProductionsAdded()
-                ).choose(player)
-        );
-        player.getGameClientHandler().sendEvent(new ActionDoneEvent("You deleted a production!"));
+        // returning a fail event if there are no productions added
+        if (!player.getProductionsAdded().isEmpty()) {
+            player.deleteProduction(
+                    new MakePlayerChoose<>(
+                            "Choose the production you want to delete",
+                            player.getProductionsAdded()
+                    ).choose(player)
+            );
+            player.getGameClientHandler().sendEvent(new ActionDoneEvent("You deleted a production!"));
+        }
+        else{
+            player.getGameClientHandler().sendEvent(new FailEvent("No productions to be deleted!"));
+        }
     }
 }
